Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    float str1, str2;
+    double str1, str2;
 
   
     scanf("%if %if", str1, str2);  // Reads two strings separated by space
